+++ content_editor/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import absolute_import, unicode_literals
 
 
-VERSION = (1, 3, 1)
+VERSION = (1, 3, 2)
 __version__ = ".".join(map(str, VERSION))